+++ .gitignore
@@ -19,6 +19,7 @@
 .DS_Store
 *.pem
 .idea
+.vscode
 
 # debug
 npm-debug.log*
