+++ app/(landing)/Prizes.tsx
@@ -1,4 +1,4 @@
-import Section from './Section';
+import Section from '../../components/Section';
 
 
 export default function Prizes() {
