+++ app/(landing)/Schedule.tsx
@@ -1,5 +1,5 @@
 import {ReactNode} from 'react';
-import Section from './Section';
+import Section from '../../components/Section';
 
 
 export default function Schedule() {
