+++ app/(landing)/Sponsors.tsx
@@ -1,4 +1,4 @@
-import Section from './Section';
+import Section from '../../components/Section';
 import Sponsor from './Sponsor';
