+++ app/(landing)/head.tsx
@@ -1,0 +1,8 @@
+export default function Head() {
+    return (
+        <>
+            <title>GunnHacks 9.0</title>
+            <meta name="description" content="Make, Build, Create & Learn. It’s GunnHacks 9.0, Gunn’s 24‑hour high school hackathon." />
+        </>
+    )
+}
