+++ next.config.js
@@ -2,6 +2,9 @@
 const nextConfig = {
   reactStrictMode: true,
   swcMinify: true,
+  experimental: {
+    appDir: true
+  }
 }
 
-module.exports = nextConfig
+module.exports = nextConfig;
