--- pages/_app.tsx
@@ -1,8 +1,0 @@
-import {AppProps} from 'next/app';
-
-import '../styles/globals.css'
-
-
-export default function App({ Component, pageProps }: AppProps) {
-    return <Component {...pageProps} />;
-}
